Move HTTPGetHandler to http_handler package
