+++ cmd/main.go
@@ -7,6 +7,7 @@ import (
 	"github.com/keptn-contrib/dynatrace-service/internal/credentials"
 	"github.com/keptn-contrib/dynatrace-service/internal/env"
 	"github.com/keptn-contrib/dynatrace-service/internal/event_handler"
+	"github.com/keptn-contrib/dynatrace-service/internal/http_handler"
 	"github.com/keptn-contrib/dynatrace-service/internal/onboard"
 
 	log "github.com/sirupsen/logrus"
@@ -46,7 +47,7 @@ func _main(args []string, envCfg envConfig) int {
 	ctx = cloudevents.WithEncodingStructured(ctx)
 
 	log.WithFields(log.Fields{"port": envCfg.Port, "path": envCfg.Path}).Debug("Initializing cloudevents client")
-	p, err := cloudevents.NewHTTP(cloudevents.WithPath(envCfg.Path), cloudevents.WithPort(envCfg.Port), cloudevents.WithGetHandlerFunc(HTTPGetHandler))
+	p, err := cloudevents.NewHTTP(cloudevents.WithPath(envCfg.Path), cloudevents.WithPort(envCfg.Port), cloudevents.WithGetHandlerFunc(http_handler.HTTPGetHandler))
 	if err != nil {
 		log.WithError(err).Fatal("Failed to create client")
 	}
