+++ internal/http_handler/health_handler.go
@@ -1,4 +1,4 @@
-package main
+package http_handler
 
 import (
 	"net/http"
