+++ internal/http_handler/http_get_handler.go
@@ -1,4 +1,4 @@
-package main
+package http_handler
 
 import (
 	"net/http"
