+++ internal/http_handler/not_found_handler.go
@@ -1,4 +1,4 @@
-package main
+package http_handler
 
 import (
 	"encoding/json"
